worker/requirements.txt
pydantic>=2.8,<3
pydantic-settings>=2.3,<3
motor>=3.6,<4
pymongo>=4.7,<5
google-cloud-storage>=2.18,<3
google-auth>=2.33,<3
ultralytics>=8.3.0
pillow>=10.4,<11
tenacity>=9.0,<10
structlog>=24.1.0
tqdm>=4.66,<5
